requirements.txt
# Web框架
Flask==2.2.2
Flask-CORS==4.0.0
Jinja2==3.1.2

# 数据库驱动
pymongo==4.3.3
redis==4.6.0

# 数据处理
pandas==1.5.3
numpy==1.24.2
openpyxl==3.1.2

# 机器学习和相似度计算
scikit-learn==1.2.1
rapidfuzz==3.0.0

# 中文处理
jieba==0.42.1
pypinyin==0.47.1

# 异步任务处理
celery==5.3.2
kombu==5.3.2

# 工具库
PyYAML==6.0
python-dotenv==1.0.0
requests==2.28.2
Werkzeug==2.2.2

# 日志和监控
loguru==0.7.2

# 开发工具
pytest==7.4.2
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0